Update comments in deploy_api_to_server.yml for clarity
Updated comments throughout the deploy_api_to_server.yml file for clarity and consistency. Changes include specifying the use of Ubuntu for building and file transfer, clarifying the purpose of listing repository files for debugging, detailing the setup of the .NET SDK, and specifying the use of the solution file for restoring dependencies. Additional updates include clarifying the build mode as Release, detailing the publish folder, and improving descriptions for zipping files, pre-creating target folders, installing sshpass, unzipping files with PowerShell, optional cleanup of zip files, and restarting the IIS website on the Windows EC2 instance.