Add IIS service control steps in deployment script
Added steps to stop and start the IIS service (W3SVC) on the EC2
Windows instance in the `deploy_api_to_server.yml` file. Specifically,
a step to stop the IIS service was added before clearing the target
folder, and a step to start the IIS service was added before starting
the IIS application pool.