Rename API directory and add Currency property
Renamed the directory from `API` to `Api` and updated all related configuration files to reflect this change. Updated `NuGetToolVersion` in `API.csproj.nuget.g.props` from `6.12.3` to `6.12.2`. Added a new `Currency` property to `PropertyDto` and `LandLordProperty` classes. Introduced a new `Tenant` class. Added a migration to handle the addition of the `Currency` column to the `LandLordProperties` table. Updated various assembly info and cache files with new values and hashes.